passo codice

+++ constants.h
@@ -9,4 +9,4 @@
 #define MAX_VELOCITY 0.5 // meters/seconds (approssimazione ricavata contanto i giri di ruota in 5 secondi con pwm al massimo)
 #define MAX_TICK_FREQ (MAX_VELOCITY / WHEEL_CIRCUMFERENCE * TICKS_PER_REV)
 #define MAX_TICK_DURATION 1 / MAX_TICK_FREQ
-#define SECS_TO_MICROS pow(10, 6)
+#define SECS_TO_MICROS powf(10, 6)
